requirements.txt
alembic==1.0.8
arrow==0.13.1
blinker==1.4
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dateparser==0.7.1
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud==0.34.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
moment==0.8.2
oauth2client==4.1.3
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
regex==2019.2.21
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.0
times==0.7
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
